refactor: remove not needed file
