+++ inc/utils_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils_bonus.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+        */
+/*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/29 11:46:04 by u413q             #+#    #+#             */
-/*   Updated: 2024/01/19 12:39:04 by sqiu             ###   ########.fr       */
+/*   Updated: 2024/01/21 13:05:04 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,6 @@
 # include <math.h>
 # include <stdbool.h>
 # include <stdlib.h>		// required for rand()
-# include <time.h>
 
 //home-grown libs
 # include "vec3.h"
@@ -33,7 +32,7 @@
 
 /**
  * @brief Interval of accepted values
- * 
+ *
  * @param min	Minimum of interval
  * @param max 	Maximum of interval
  */
@@ -45,16 +44,16 @@ typedef struct s_interval
 
 /**
  * @brief Contains factors to solve linear and quadratic equations.
- * 
+ *
  * ax² + bx + c = 0
- * 
+ *
  * x = (-b +/- sqr(b² - 4ac)) / 2a
  * Here we use d instead of x as variable to represent the distance of the
  * ray traveled.
  * @param a		Factor of quadratic component
  * @param b		Factor of linear component
  * @param c		Constant
- * @param d1	First solution 
+ * @param d1	First solution
  * @param d2	Second solution
  */
 typedef struct s_equation
@@ -68,7 +67,7 @@ typedef struct s_equation
 
 /**
  * @brief Quaternion definition with
- * 
+ *
  * @param scalar	Real component
  * @param vector	Imaginary component consisting of three values/directions
  */
@@ -82,7 +81,7 @@ typedef struct s_quaternion
 
 /**
  * @brief Transforms degrees into radian
- * 
+ *
  * @param degrees 	Angle in degrees
  * @return float 	Angle in radians
  */
@@ -98,7 +97,7 @@ float			ft_radian_to_degree(float radians);
 
 /**
  * @brief Determines if value is inside the interval
- * 
+ *
  * @param x 			Value
  * @param interval 		Interval with min and max
  * @return true 		Value is inside the interval
@@ -108,7 +107,7 @@ bool			ft_contains(float x, t_interval interval);
 
 /**
  * @brief Determines if value is surrounded by the interval
- * 
+ *
  * @param x 			Value
  * @param interval 		Interval with min and max
  * @return true 		Value is surrounded by the interval
@@ -118,12 +117,12 @@ bool			ft_surrounds(float x, t_interval interval);
 
 /**
  * @brief Solves the given equation
- * 
+ *
  * If the parameter a = 0, the equation is linear and the sole
  * solution is calculated and stored in d1.
  * Otherwise the equation is quadratic and dependent on the
  * discriminant real solutions exist or not:
- * 
+ *
  * Discriminant < 0: no real solution, return -1
  * Discriminant = 0: one solution, return 0
  * Discriminant > 0: two solutions, return > 0
@@ -135,7 +134,7 @@ float			ft_solve(t_equation *eq);
 
 /**
  * @brief Checks if two floats are (almost) equal
- * 
+ *
  * If their absolute difference is below threshold EPSILON,
  * then they are declared equal
  * @param one 		First float
@@ -147,52 +146,52 @@ bool			ft_nearly_equal_flt(float one, float two);
 
 /**
  * @brief Return a random float
- * 
- * @return float 
+ *
+ * @return float
  */
 float			ft_random_float(void);
 
 /**
  * @brief Return a random float within min and max
- * 
+ *
  * @param min 		Lower bounder
  * @param max 		Upper bounder
- * @return float 
+ * @return float
  */
 float			ft_random_float_within(float min, float max);
 
 /**
  * @brief Ensure that the returned float lies within the interval
- * 
+ *
  * @param x 			Float given
  * @param interval 		Interval with min and max value
- * @return float 
+ * @return float
  */
 float			ft_clamp(float x, t_interval interval);
 
 /**
  * @brief Performs quaternion multiplication
- * 
- * @param q1	First quaternion 
+ *
+ * @param q1	First quaternion
  * @param q2 	Second quaternion
- * @return t_quaternion 
+ * @return t_quaternion
  */
 t_quaternion	ft_quaternion_mult(t_quaternion q1, t_quaternion q2);
 
 /**
  * @brief Performs a rotation of a given vector around an axis by
  * a given angle
- * 
+ *
  * The vector is transformed into a quaternion p. A quaternion q is
  * created with the axis and the given angle. The inverse of q is
- * calculated by setting the vector of q_inv as the vector of q 
- * scaled by -1. 
+ * calculated by setting the vector of q_inv as the vector of q
+ * scaled by -1.
  * The rotation is effected by multiplying q with p, and thereafter
  * the result with q_inv (lefthand multiplication).
  * @param vec 		Vector to be rotated
  * @param axis 		Axis around which to rotate
  * @param angle 	Amount of rotation
- * @return t_vec3 
+ * @return t_vec3
  */
 t_vec3			ft_quaternion_rot(t_vec3 vec, t_vec3 axis, float angle);
