+++ src/colour.c
@@ -6,7 +6,7 @@
 /*   By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/27 11:30:57 by u413q             #+#    #+#             */
-/*   Updated: 2024/01/19 15:16:13 by gwolf            ###   ########.fr       */
+/*   Updated: 2024/01/21 13:07:07 by gwolf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,4 +30,28 @@ t_colour	ft_hadamard_colour(t_colour c1, t_colour c2)
 		.b = ft_clamp(c1.b * c2.b, interval),
 		.g = ft_clamp(c1.g * c2.g, interval)
 	});
+}
+
+uint32_t	ft_convert_colour2int(t_colour colour)
+{
+	uint8_t	a;
+	uint8_t	r;
+	uint8_t	g;
+	uint8_t	b;
+
+	a = 0;
+	r = (uint8_t)(255.9 * colour.r);
+	g = (uint8_t)(255.9 * colour.g);
+	b = (uint8_t)(255.9 * colour.b);
+	return ((a << 24) | (r << 16) | (g << 8) | b);
+}
+
+t_colour	ft_convert_int2colour(uint32_t colour)
+{
+	t_colour	ret;
+
+	ret.r = colour >> 16 & 0xFF;
+	ret.g = colour >> 8 & 0xFF;
+	ret.b = colour & 0xFF;
+	return (ret);
 }
